requirements.txt
openai
flask
flask-cors
PyPDF2
python-docx
openpyxl
Pillow
pytesseract
